+++ lib/screens/settings/settings_screen.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class SettingScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      
+    );
+  }
+}
